Secret key input: Accept longer base keys
The hash operation requires an exactly 16-byte key, but assuming the
provided key is generated randomly (RFC 4086,
https://tools.ietf.org/html/rfc4086) and has sufficient entropy to be
used by as a cryptographic key, it is possible to consistently select 16
bytes from the input material to use as a hashing key to generate a MAC.

Notably, increasing key length does not improve security for this
system, as the limiting factor for 'cracking a keycode' is how much
information of the output MAC is placed into the keycode. For example,
if a keycode only contains 3 'digits (0-9)' of MAC content, an attacker is
likely to guess a correct keycode for the unit after only 500 attempts.

In contrast, a 16-byte/128 bit key takes 2^128 operations to guess every
possible key. Assuming the power of the entire hashrate of the bitcoin
network were put behind this problem, it should take 2,158,000,000,000 years to
guess a secret key. In other words, over 100x the age of the universe.
(ref:
https://crypto.stackexchange.com/questions/48667/how-long-would-it-take-to-brute-force-an-aes-128-key)

While 32-byte keys may be useful for other purposes (providing more
input material for further key derivation steps), they do not inherently
provide 'more security' than an 128-bit key.