feat: Introduce obscured origin commands
Using the same obscuring logic used by standard activation keycodes,
obscure the non-MAC digits of origin commands. As a reminder, the last 6
digits are MAC digits, and after the origin command is generated,
the digits '81' are appended in front if conveying as a Nexus Keycode
Passthrough command (passthrough ID = 8, subtype ID 1 = "Nexus Channel
Origin Command").

If not conveying the origin command via Nexus Keycode, the result of
`to_digits` can be passed to a device directly.